refactor: ListItemView와 SearchView TCA 아키텍쳐 적용
ListItemView -> ListItemFeature 리듀서 사용

SearchView -> SearchFeature 리듀서 사용
