+++ CleanArea/CleanArea/iOS/Main/Components/LoadingCircleView.swift
